requirements.txt
fastapi
uvicorn[standard]
pymssql
python-dotenv
pydantic
taskipy
ruff
